+++ app/schemas/post_create_dto.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 
-class Post:
+class PostCreateDTO:
     def __init__(self, id: str | None, title: str, content: str, member_id: int, board_id: int,
                  created_at: datetime):
         self._validate_title(title=title)
