+++ app/services/post_service.py
@@ -1,9 +1,9 @@
 from datetime import datetime
 
-from app.models.post import Post
+from app.schemas.post_create_dto import PostCreateDTO
 
 
-def create_post(title: str, content: str, member_id: int, board_id: int) -> Post:
+def create_post(title: str, content: str, member_id: int, board_id: int) -> PostCreateDTO:
     """
     게시글을 작성합니다.
 
@@ -13,6 +13,6 @@ def create_post(title: str, content: str, member_id: int, board_id: int) -> Post
     @:param board_id 게시판의 식별자
     @:return created_post 저장된 게시글
     """
-    post = Post(id=None, title=title, content=content, member_id=member_id, board_id=board_id, created_at=datetime.now())
+    post = PostCreateDTO(id=None, title=title, content=content, member_id=member_id, board_id=board_id, created_at=datetime.now())
     # created_post = 게시글 저장
     # return created_post
